Frontload Spectrum Components to fix tooltip behavior
